requirements.txt
# Core ML + Torch stack (with RTX 5090 nightly builds)
torch==2.8.0.dev20250603+cu128
torchvision==0.23.0.dev20250604+cu128
torchaudio==2.8.0.dev20250604+cu128

# Core ML libraries
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.3
pandas==2.2.3

# Image processing
opencv-python==4.11.0.86
pillow==11.2.1
ttach==0.0.3

# ONNX export & inference
onnx==1.18.0
onnxruntime==1.22.0

# Progress bars
tqdm==4.67.1

# Logging / utils
PyYAML==6.0.2

# GradCAM (compatible with Torch Nightly)
git+https://github.com/jacobgil/pytorch-grad-cam.git@master

# Local package (editable, for development)
-e .
